Enable use of Keras graph functions with iterators and datasets during eager execution. This is done by routing `fit`/etc to the methods `fit_generator`/etc when passed an eager dataset or eager iterator.
PiperOrigin-RevId: 222136116
