+++ tensorflow/python/keras/engine/training_utils.py
@@ -140,7 +140,7 @@ def convert_to_iterator(x=None,
   if isinstance(x, iterator_ops.EagerIterator):
     if steps_per_epoch is None:
       raise ValueError('You must specify the number of steps (number of batches'
-                       'to draw from the iterator).')
+                       ' to draw from the iterator).')
     return x, steps_per_epoch
 
   if not _nested_any(sample_weights, lambda x: x is None):
@@ -155,7 +155,7 @@ def convert_to_iterator(x=None,
   data = _convert_lists_to_tuples(data)
   if steps_per_epoch is None and batch_size is not None:
     num_samples = _get_batch_axis_size(data)
-    steps_per_epoch = int(math.ceil(num_samples / batch_size))
+    steps_per_epoch = int(math.ceil(num_samples / int(batch_size)))
 
   if steps_per_epoch is None:
     alternative_arg_name = (
